Examples: Moved example into 'composition' folder, and improved it
